Add files via upload
This file includes the code for filtering the congressional_speech_df into equal pay speeches, the code for cleaning the text, and the code used in Figure 5.1 of the thesis. It also includes code for a sentiment analysis of each speech and Latent Dirichlet Allocation (LDA) topic modeling on the speech dataset. While my project did not ultimately include the results of these analyses in the final version, I’ve included this code here in case it’s useful to anyone who wants to build on or extend this work in the future.